+++ dummy_file_generator/rowdatagenerator.py
@@ -2,13 +2,22 @@
 from random import randint
 
 from dummy_file_generator.exceptions import DummyFileGeneratorException
-from dummy_file_generator.utils import whitespace_generator
+
+
+def whitespace_generator(i) -> str:
+    """
+    whitespaces generator function
+    :param i:
+    :return:
+    """
+    return int(i) * ' '
 
 
 class CsvRowDataGenerator:
     """
     csv row data generator implementation
     """
+
     def __init__(self, data_files_contents, columns):
         self.data_files_contents = data_files_contents
         self.columns = columns
@@ -32,10 +41,10 @@ class CsvRowDataGenerator:
             try:
                 _column_values_list, _column_values_list_item_count = \
                     self.data_files_contents[column['datafile']]
-            except AttributeError as attr_err:
+            except KeyError as key_err:
                 raise DummyFileGeneratorException(f'Cannot find corresponding data_file for '
                                                   f'column {column.get("column_name")}, '
-                                                  f'Attribute Error: {attr_err}')
+                                                  f'Key Error: {key_err}')
 
             value = _column_values_list[randint(0, _column_values_list_item_count - 1)]
             row.append(value)
@@ -47,6 +56,7 @@ class FlatRowDataGenerator:
     """
     flat row data generator implementation
     """
+
     def __init__(self, data_files_contents, columns):
         self.data_files_contents = data_files_contents
         self.columns = columns
@@ -62,7 +72,7 @@ class FlatRowDataGenerator:
 
         for _column_values_list, _column_length in zip(self.column_names,
                                                        self.column_lengths):
-            _header_row.append(str(_column_values_list) +
+            _header_row.append(_column_values_list +
                                whitespace_generator(_column_length - len(_column_values_list)))
 
         header_row = "".join(_header_row)
@@ -81,12 +91,12 @@ class FlatRowDataGenerator:
                 _column_values_list, _column_values_list_item_count = \
                     self.data_files_contents[column['datafile']]
                 _whitespace_count = column.get('column_len')
-            except AttributeError as attr_err:
+            except KeyError as key_err:
                 raise DummyFileGeneratorException(f'Cannot find corresponding data_file for '
                                                   f'column {column.get("column_name")}, '
-                                                  f'Attribute Error: {attr_err}')
+                                                  f'Key Error: {key_err}')
             value = _column_values_list[randint(0, _column_values_list_item_count - 1)]
-            row.append(str(value) + whitespace_generator(_whitespace_count - len(value)))
+            row.append(value + whitespace_generator(_whitespace_count - len(value)))
 
         row = "".join(row)
 
@@ -97,6 +107,7 @@ class RowDataGenerator:
     """
     row data generator factory
     """
+
     def __init__(self, file_type, data_files_contents, columns):
         _mapped_generator_class = {
             "csv": CsvRowDataGenerator,
